refactor: integrate template-primitives conformance into existing per-adapter tests (#1194 review)
Each adapter package already has a single test file that bundles the
JSX conformance call (`runJSXConformanceTests`) with adapter-specific
test groups (`describe(...)` blocks) — `hono-adapter.test.ts`,
`go-template-adapter.test.ts`. The previous commit's standalone
`template-primitives-conformance.test.ts` per adapter splintered the
suite list without reason; merge them back into the canonical adapter
test files.

Adapter authors now have one place per package to find every shared
conformance call and adapter-specific test, mirroring the existing
`runJSXConformanceTests` pattern.

No behaviour change. Test counts unchanged from the previous commit
(hono: +4 conformance, go-template: +4 skip, adapter-tests: -8 from
the central deletion).
