+++ packages/adapter-go-template/src/__tests__/go-template-adapter.test.ts
@@ -6,7 +6,14 @@
 
 import { describe, test, expect } from 'bun:test'
 import { GoTemplateAdapter } from '../adapter/go-template-adapter'
-import { runJSXConformanceTests } from '@barefootjs/adapter-tests'
+import {
+  runJSXConformanceTests,
+  runConformanceSuite,
+  templatePrimitiveCases,
+  runTemplatePrimitiveCase,
+  TemplatePrimitiveCaseId,
+  type TemplatePrimitiveInput,
+} from '@barefootjs/adapter-tests'
 import { renderGoTemplateComponent, GoNotAvailableError } from '@barefootjs/go-template/test-render'
 import { compileJSX, type ComponentIR } from '@barefootjs/jsx'
 
@@ -50,6 +57,31 @@ runJSXConformanceTests({
     }
     return false
   },
+})
+
+// =============================================================================
+// Template-Primitive Conformance (#1187 phase 3)
+// =============================================================================
+
+// Go's template runtime is the html/template engine — it can render
+// only callees the adapter explicitly maps to a Go template function
+// via `templatePrimitives`. None mapped yet (#1188 will land them);
+// every positive-inlining case stays skipped until that PR.
+runConformanceSuite<typeof TemplatePrimitiveCaseId[keyof typeof TemplatePrimitiveCaseId], TemplatePrimitiveInput, string>({
+  name: 'template primitives conformance',
+  issue: '#1187 phase 3',
+  adapter: {
+    name: 'go-template',
+    factory: () => new GoTemplateAdapter(),
+    skip: new Set([
+      TemplatePrimitiveCaseId.JSON_STRINGIFY_VIA_CONST,
+      TemplatePrimitiveCaseId.MATH_FLOOR_VIA_CONST,
+      TemplatePrimitiveCaseId.USER_IMPORT_VIA_CONST,
+      TemplatePrimitiveCaseId.NO_DOUBLE_REWRITE_OF_PROPS_OBJECT,
+    ]),
+  },
+  cases: templatePrimitiveCases,
+  run: runTemplatePrimitiveCase,
 })
 
 // =============================================================================
